[#29249] docdb: (test-only) fix analyze test w/ vlogging enabled
Summary:
When the hardcoded constant `kEnableVlogsDuringAnalyze` is set to true, the `PgAnalyzeTest` tests restart the postmaster. The machinery they use to do this doesn't work after the introduction of the YSQL lease in 53146f9da4877244f931761920a5159a5b8da412.

This diff uses the tserver callback instead to restart pg, which works with the lease enabled.
Jira: DB-19024

Test Plan:
apply diff:
```
-  static constexpr auto kEnableVlogsDuringAnalyze = false;
+  static constexpr auto kEnableVlogsDuringAnalyze = true;
```

run:
```
./yb_build.sh release --cxx-test pg_analyze-test --cxx-test-filter-re pg_analyze-test --gtest_filter PgAnalyzeTest.AnalyzeSamplingColocated --test-timeout-sec 1800
```

Reviewers: timur

Reviewed By: timur

Subscribers: svc_phabricator, yql, rthallam, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D48063