[BACKPORT 2025.2][#29315] YSQL: Fix TestRollbackToSavepointWithDropCreateTableSameName in TSAN/ASAN
Summary:
The test TestRollbackToSavepointWithDropCreateTableSameName has been quite flaky especially in slower builds such as TSAN and ASAN. This
revision fixes the test.

The issue happens when a table gets deleted as part of rollback to savepoint operation. When we have to drop a table in such scenarios, we don't wait
for the table deletion to finish. Instead, we assume it'll be successful once the table has been marked DELETING. See the call to
`RemoveDdlRollbackToSubTxnState` in the YsqlDdlTxnDropTableHelper function. This is safe to do because the deleted table can only be seen by the
current transaction since it must have been created in the same txn block.

In the test, we trigger the rollback to savepoint operation and immediately check for the deletion of the ysql_ddl_txn_verifier_state from the table.
The verifier state is only deleted at the later stages of the table deletion i.e. after the tablet deletion. As a result, on slower builds, we end up
making the assertion sooner than the deletion completion.

Fixed the test by introducing a wait before we assert for the cleanup of verifier state.

Original commit: 390cce182c23367681d7eafe1b32b27c6dddf01d / D48195

Test Plan:
Jenkins: test regex: .*PgDdlSavepointMiniClusterTest.*

./yb_build.sh tsan --cxx-test pgwrapper_pg_ddl_transaction-test --gtest_filter
bool/PgDdlSavepointMiniClusterTest.TestRollbackToSavepointWithDropCreateTableSameName/0 -n 10

./yb_build.sh tsan --cxx-test pgwrapper_pg_ddl_transaction-test --gtest_filter
bool/PgDdlSavepointMiniClusterTest.TestRollbackToSavepointWithDropCreateTableSameName/1 -n 10

Reviewers: pjain, myang

Reviewed By: myang

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D48211