[#29260] YSQL: incremental refresh of relcache init file
Summary:
With invalidation messages, we can now try to revalidate a stale relcache init file if its all of
its contents except the stored catalog version remain the same. Let's say initially we have an
up-to-date file with catalog version 30 stored in the relcache init file. After a few DDL statements, the
catalog version changed from 30 to 35. The on disk relcache is now considered stale based upon the
following check:

```
        /*
         * If we already have a newer cache version (e.g. from reading the
         * shared init file) or master has newer catalog version then this file is too old.
         */
        if (YbGetCatalogCacheVersion() > ybc_stored_cache_version)
        {
            yb_stale_version = true;
            unlink_initfile(initfilename, ERROR);
            goto read_failed;
        }
```
Though this is safe to do, it is possible that from 30 to 35, those invalidation messages do not
touch any relations stored in the relcache init file at all. In other words, after we have
successfully applied all those messages, we would generate a new relcache init file which has
exactly the same contents as before, except the stored catalog version is changed from 30 to 35.

Instead of rebuilding a new relcache file by triggering a new internal connection which performs
catalog preloading that takes time and memory, we can send a local RPC to the tserver to fetch the
list of invalidation messages from 30 to 35. When we scan through the invalidation messages to see
if any one of them could touch one of the relations in the relcache init file. If there isn't such a
message, we can simply revalidate the relcache init file which will be more efficient.

This diff changes the way a new connection handles a stale cache file. The new backend now calls
`YbTryRevalidateRelcacheFile` to check if an incremental update is possible. If successful, the
backend simply changes the stored catalog version from 30 to 35 and go ahead to load the relcache
init file as usual. After loading the on-disk relcache entries into memory, the backend simply calls
`write_relcache_init_file` to write the same relcache with version 35.

NOTES:

1. This diff does not affect active connections which does not rewrite the relcache init
file (even when it performs a full refresh, it only rebuild relcache in memory). Only new
connections are affected because they are responsible for rewriting the file.

2. We rewrite the entire file contents, instead of only updating the catalog version 30 to 35. The
reason is that when loading the relcache init file, the file is opened in read only mode. In order
to update 30 to 35, we would need to close the file and then reopen it in read/write mode. It is
possible that between closing and reopen the file, another backend has just rewrote the file to have
version 36, and it will be wrong if this backend overwrites that with 35. Even though we can add
more code logic to resolve that race condition, I feel that it is just simple enough to call
`write_relcache_init_file` to rewrite the entire file (which isn't too big) instead of attempting
micro-optimization to update only the version.

3. Due to garbage collection of invalidation messages in the tserver, it is still possible that all
the active connections have moved beyond version 31, so the invalidation messages of version 31 has
already been GC'ed. In this case incremental refresh of the file will fail and we fall back to the
existing behavior. There are also other reasons for `YbTryRevalidateRelcacheFile` to fail and we
fall back to the existing behavior: trigger an internal connection to rebuild the file via catalog
preloading. The unit test NewConnectionRelCachePreloadTest is updated to continue to test the
existing behavior by bumping up the catalog version only without an invalidation message list.

Test Plan: jenkins

Reviewers: kfranz, sanketh

Reviewed By: kfranz

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D48288