[PLAT-18636] Fix missing health check metrics
Summary:
We have a logic which exports the metrics form health check failed nodes count.
This logic was implemented long time ago, when health check script was on YBA node.
And it had a logic to determine between real check failure and node being completely down.
The logic was dependent on the checks order in the response. Order changed at some point later - and
metrics calculation was broken with that.
Now, once we execute the check on the DB node - this is not needed anymore. AS completely dead node
will only have one item in the response saying that check failed completely.
Removed this logic + also reordered checks to make the report more reasonable (important check first)

Test Plan: Tested manually by creating 2 universes - rf1 co-located and rf1 dedicated master.

Reviewers: skurapati

Reviewed By: skurapati

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D48272