[#29362] YSQL: Add GUC to allow toggling of FK correctness fix
Summary:
As part of f225ca435, fk triggers started using a fresh snapshot instead
of the transaction snapshot. This was required to fix correctness issue
described in that commit. For this fix, a prequisite commit (18eb1a9fc5) was
required to integrate PG's snapshot managament with the analogous
read_time_serial_no management in pggate. This integration was already in place
for READ COMMITTED isolation, the prereq commit extended it to
REPEATABLE READ isolation level.

This revision introduces a GUC to allow a user to fallback to the mode
before both commits. Also, the fallback would work even if we had only
reverted to a state without 18eb1a9fc5 because without this pre-req
commit, the new `GetLatestSnapshot` call that is made as part of
`f225ca435` would be a no-op. However, to be cautious, the GUC
would revert the change in `f225ca435` too (to avoid the PG
state change as well).

Use of the GUC is not recommended and needs extreme caution. It
should be used only with a rolling restart after adding to
ysql_pg_conf_csv. It could also be used in a session in between
transactions (i.e. not in the middle of an active transaction), but that
option should be exercised with caution and discussion after evaluating
the workload.

NOTE: this is a 2025.1 only commit. The fallback option is not added to
future releases to avoid complexity in code that is actively being
updated. Also, future releases will have much more soak time for the
changes.

Test Plan:
Jenkins: rebase: 2025.1

Manually tried the new fallback GUC and executed the test case mentioned in the summary of D46367. The bug resurfaces (i.e., the old behaviour is restored.

Reviewers: smishra, kramanathan, dmitry

Reviewed By: kramanathan

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D48305