[BACKPORT 2.29.0.1081] [PLAT-18986] Use the right userIntent for YNP provisioning
Summary:
During YNP provisioning, we have to setup a configuration file with
node and provider level details. We use the cluster, which the node
is a part of, to determine its details and retrieve provider level
details (e.g. the image bundle associated with the provider).

In YBM, we create seperate clusters for the primary cluster and the
read-replica cluster with independent providers. However, we ended
up using the primary cluster's unserIntent to retrieve the provider
level details which is incorrect. We need to use the read-replica
cluster's provider to retrieve this information. This change fixes
the issue by retrieving the right cluster which the node is a part
of for getting the provider level details.

In addition to the above bug fix, this change addresses two other
prior issues:
    - First, we need to configure NICs before running YNP based
       provisioning because configure NICs will reboot the node
       which may clean-up the temporary config created in tmp space
    - Second, we need to re-throw any exceptions caught while
       getting provisioning parameters for YNP based provisioning
    -  Third, we need to now fix 3 UTs for VMImageUpgrade tests
        which were failing because YNP based provisioning now
        throws an exception when it fails to generate the node
        configuration

Test Plan: Ran local UTs and create universe with primary and read replica clusters

Reviewers: anijhawan, nsingh, rpendse

Reviewed By: rpendse

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D48318