[#29389] YSQL: Acquire object locks on relation(s) for yb_index_check
Summary:
As part of `yb_index_check`, YB reads both the base table and the index table as of time and ensures that the index is consistent.

The tests `testPgRegressYbIndexCheck` and `testPgRegressYbIndexCheckSingleSnapshot` fail when object locking is enabled, as they trigger the debug sanity check where we assert that some locks are taken before a read/write is performed. This is because the relations weren't being open using `relation_open`, which acquires the relevant object locks.

This revision addresses the issue by explicitly opening the relation (both the base and the index) in `AccessShareLock` mode, as a result of which the relevant object locks are acquired.

Test Plan:
Jenkins

Enabled object locking feature and ran the following tests
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressYbIndexCheckSingleSnapshot#testPgRegressYbIndexCheckSingleSnapshot'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressYbIndexCheck#testPgRegressYbIndexCheck'
```

Reviewers: amitanand, rthallam, aagrawal, #db-approvers

Reviewed By: amitanand, aagrawal

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D48378