[BACKPORT 2025.2][PLAT-18686][PLAT-18988]: Supporting backing up and restoring global role objects
Summary:
We have started to support backing up and restoring global role objects

Backup Changes:
At backup time, have a field to backup global roles - pass a flag to useRoles to tell it to backup the roles
- Off by default

Restore Changes:
If roles were not backed up
- When trying to restore do not show "Restore Gloabl roles"

If roles were backedup in the first place
- Show option to "Restore Global roles" and have it Off by default
- if it turned ON, then show one more field "Fail restore if role already exists in target database"
    - Default will be false
    - If turned on, we need to pass useRoles and errorIfRolesExists to true in the restore API
It does not make sense to give the option for "error if roles exists" in the UI. the backup will include default roles (such as postgres. Because its default, its auto created for new universes and will exist there too.

Original commit
1. d946d50/D47655
2. 44dea23/D48380

Test Plan:
Please refer to the video
NOTE: error if roles exists is not shown in the UI anymore, just that video is slightly older
{F406077}

Reviewers: jmak, kkannan, dshubin, vkumar

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D48451