[#28670] docdb: use recommended signals to stop postmaster process
Summary:
the ysql lease work switched to using SIGKILL to stop postmaster instead of SIGINT.
This is not recommended:
https://www.postgresql.org/docs/current/server-shutdown.html

This diff changes the signal used to stop postmaster.
When the lease is lost, we now use SIGQUIT which corresponds to PG's immediate shutdown mode.
When the tserver itself is shutting down, we use SIGINT which corresponds to PG's fast shutdown mode.

This also restores the semantics introduced by 002013127895d49a78d23bf5265461118d637ff5.

Test Plan: existing tests

Reviewers: bkolagani, kramanathan

Reviewed By: kramanathan

Subscribers: yql, timur, ybase, rthallam

Differential Revision: https://phorge.dev.yugabyte.com/D48404