[BACKPORT 2025.2][#29389] YSQL: Acquire object locks on relation(s) for yb_index_check
Summary:
As part of `yb_index_check`, YB reads both the base table and the index table as of time and ensures that the index is consistent.

The tests `testPgRegressYbIndexCheck` and `testPgRegressYbIndexCheckSingleSnapshot` fail when object locking is enabled, as they trigger the debug sanity check where we assert that some locks are taken before a read/write is performed. This is because the relations weren't being open using `relation_open`, which acquires the relevant object locks.

This revision addresses the issue by explicitly opening the relation (both the base and the index) in `AccessShareLock` mode, as a result of which the relevant object locks are acquired.

Original commit: 59c8dab1e676dd4e90e1c0afa06ea1194e772575 / D48378

Test Plan:
Jenkins

Enabled object locking feature and ran the following tests
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressYbIndexCheckSingleSnapshot#testPgRegressYbIndexCheckSingleSnapshot'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressYbIndexCheck#testPgRegressYbIndexCheck'
```

Reviewers: #db-approvers, amitanand, rthallam, aagrawal

Reviewed By: rthallam

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D48413