[#29303] YSQL: Don't use od_write in od_frontend_remote_client to avoid TCP deadlock
Summary:
`od_frontend_remote_client` is called from the frontend client's coroutine and shouldn't use any
blocking calls. When `od_write` is called, it will block if the socket's write buffer is full. When this
happens, then further calls can't be made until the write is able to complete. Now suppose if the PG
backend's write buffer is also full and it is blocked on Connection Manager reading the results
before it can continue with processing more queries. However, CM is blocked on write and will never
attempt to make a read. This results in a deadlock where both CM and backend are waiting on each
other.

The solution is to use an asynchronous write -- save the data in `relay`'s `iov` buffer. This is the
same mechanism with how a relay works, and so also makes more sense to do from the design of the
codebase.

We are following the same fix that upstream has done, but not cherry picking the entire commit since
it contains some changes we don't want to downmerge. See the upstream commit:
https://github.com/yandex/odyssey/commit/ff5b6563108d26fa12967bef249c054f9fe5b69b#diff-d23e9ec6d450d0487f0e95425e6eaea8e76945e520a720a69cf0efd2cc0403dcR1188

Similar to what is done in the above commit, we also remove the logic of the variable `forwarded`. It
was added since some packets were getting written directly using `od_write` but other packets were
getting sent asynchronously and so the packets were getting reordered. See the upstream issue
https://github.com/yandex/odyssey/issues/459. Now that we are never doing a blocking write, that
logic is not required.

Test Plan: Jenkins: all tests

Reviewers: skumar, mkumar, asrinivasan, vikram.damle, #db-approvers

Reviewed By: skumar, mkumar, #db-approvers

Subscribers: svc_phabricator, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D48157