[#29358,#28284] YSQL: Revert "Disallow disabling atomicity GUCs with txn ddl"
Summary:
This reverts commit d893e0182317a9de4d13526d48657eb95bbf4391.

As part of d893e0182317a9de4d13526d48657eb95bbf4391 / D48297, we added safeguards to disallow disabling DDL atomicity infra GUCs when txn ddl is
enabled.

Presently, this leads to issues in scenarios when the user wants to disable both ddl atomicity and transactional ddl at the same time. Since we don't
have any guarantees on the ordering of GUC changes, the DDL atomicity GUC gets disabled first leading to the safeguard throwing an error since
transactional DDL is not yet disabled.

A few tests are failing as well due to this change. So we'll revert this and find another way to implement which requires a bit of discussion.

Test Plan: N/A

Reviewers: myang, aagrawal, sumukh.phalgaonkar

Reviewed By: sumukh.phalgaonkar

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D48496