[PLAT-18871] make SetupYNP reboot nodes before retries
Summary:
SetupYNP does not use the generally available node reboot based retry logic, so it's susceptible to failure when SSH is not responding.
This is resolved by making it a subclass of NodeTaskBase which has the onFailure hook.

Test Plan: During universe creation, temporarily stop/start a VM during node agent provisioning so that the task initially fails.

Reviewers: nsingh, svarshney

Reviewed By: nsingh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D48513