[BACKPORT 2025.1][#29292] YSQL: Fix prepared stmt on temp tables when using extended query protocol
Summary:
Consider a TEMP table created in a database connection that uses the extended query protocol.
Most database drivers offer an interface to create and managed prepared statements during the lifetime of the connection. Using JDBC as an example:
```
PreparedStatement ps = extended_conn.prepareStatement("INSERT INTO temp_test VALUES (?)");
```

This instructs JDBC to:
 - Create a prepared statement
 - That references a temporary table created earlier in the connection
 - Using the extended query protocol

However, JDBC does not issue the prepared statement request to the database until the same query has run `prepareThreshold` number of times. This is a driver property that defaults to 5.
For the first 5 executions of the query, JDBC tracks the query (on the driver side). The database (server) is not aware of the prepared statement.
Therefore, the first 5 executions of the query go through the 'Parse', 'Bind', 'Execute' and 'Sync' messages of the extended query protocol. In other words, each of the first 5 runs of a query are planned and executed.

On the 5th execution, JDBC issues the query as a "named statement" which instructs the database to cache the query's parse state (`CachedPlanSource`) on the server-side and store as a prepared statement.
Since the query's parse state has been cached, JDBC skips issuing the 'Parse' message for the 6th execution of the query, and issues only 'Bind', 'Execute' and 'Sync' messages for the named statement.
Note that although the database now has a prepared statement associated with the query, it may still choose to re-plan the query. Plan caching and named statement tracking are orthogonal concepts.

Currently, in the parse phase, the database checks if a query involves a temp table to determine whether the underlying transaction requires an XID (has temp table = needs XID).
This information is then stored in the `CachedPlanSource`, so that future invocations of the named statement may reuse it, while skipping the computation.
However, for a prepared statement, the information of whether the query involves a temp table is currently restored only when the driver issues an 'EXECUTE' query.
The extended query protocol mimics the functionality of `PREPARE .. EXECUTE` by using named statements:
 - The Parse message associates a driver-supplied name with the given query.
 - The Bind message associates the named query with execution parameters and a query plan. A plan may have been saved from a previous invocation.
 - The Execute message (not to be confused with the `EXECUTE` SQL command) runs the named query using the saved plan with the given execution parameters.

Since the temp table check is not restored, the database assumes that no temp table is involved in the transaction and does not generate an XID, leading to a crash on COMMIT.

This revision fixes the above issue by restoring the temp table check on fetching the `CachedPlan` rather than on running an `EXECUTE` query.
This allows the stored temp table check value to be restored irrespective of the connection protocol being used, and irrespective of the SQL command being used.
This revision also renames the temp table check field from `usesPostgresRel` to `yb_plan_references_pg_rel` to keep up with the latest coding standards.

**Postgres terminology**
 - Named statement - A query that has a name associated with it. This allows a driver using the extended query protocol to reference a query by name rather than by query string. On the database side, this has the benefit that the query does not need to be re-parsed.
 - `Bind message (stmt_name, param_list)` - A connection protocol message which instructs the database to associate the supplied parameters with a previously parsed statement. If a name is supplied, the database binds to a query with the given name. Else, the database binds to previous query.
 - CachedPlanSource - A database-side object that optionally associates a query with a cached plan. This object also contains other useful information such as the original query string, whether the query uses temp tables, etc
 - Prepared statement - A meta object on the database side that associates a statement name with a `CachedPlanSource`. Confusingly, this name can reference either the driver-supplied statement name or the name supplied as part of the `PREPARE stmt_name AS <query>;`

Original commit: b636e05c5066001e757c30f5d4e392a21d2b7470 / D48326

Test Plan:
Run the following test:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgPrepareExecute#testPrepareExecuteOnTempTable'
```

Reviewers: william.mckenna, fizaa

Reviewed By: fizaa

Subscribers: yql, smishra

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D48527