[#29242] YSQL: Exit process on pggate heartbeat failure during init
Summary:
D47823 / 89d841944e07217a125db64a2b1bb0457cfdb77e introduced the handling of errors in the heartbeat RPC between pggate and its local tserver.
A new backend acquires a session ID from its local tserver as part of its first heartbeat during the initialization of PgClient. This session ID is used in all future communication between this backend and the tserver.
Therefore, if a backend fails to acquire a session ID due to a failure of its first heartbeat, no meaningful work can be done by the backend and it must exit itself.
Currently, a `CHECK_OK` asserts that the first heartbeat succeeds during the initialization of PgClient. When the first heartbeat fails, the `CHECK_OK` causes the backend to abort itself, resulting in a core dump.

This revision reworks the error handling for the failure of the first heartbeat by:
 - Removing the `CHECK_OK` macro.
 - Propagating the error as a `Status` up to Postgres.
 - `FATAL`ing the process in Postgres which does not produce a core dump by default.

Transient network failures are not rare. Skipping the production of core dumps results in lower noise in QA stress tests and fewer support engagements in customer clusters (as alerts are raised when cores are produced).

Test Plan:
Run the following test:
```
./yb_build.sh --cxx-test pgwrapper_pg_mini-test --gtest-filter 'PgHeartbeatFailureTest.MockTransientHeartbeatFailure'
```

Reviewers: schandra, amartsinchyk, mihnea, jason, #db-approvers

Reviewed By: jason, #db-approvers

Subscribers: rthallam, svc_phabricator, dmitry, ybase, smishra, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D48477