[#29373] YSQL: Fix txn self abort for index + base table deletion in ROLLBACK TO SAVEPOINT
Summary:
As part of 913988a7d246f8a191b81ab0ca383f609d11adc0 / D47796, we fixed the transaction self abort when a table gets deleted due to rollback to savepoint (sub-transaction) operation. See the summary for more details.

We missed one case in that fix. The case:
1. Table being deleted is an index
2. Base table is also deleted as part of the rollback to savepoint operation

The issue is that for such indexes, we skip the deletion operation since they'll be deleted when the base table is deleted. As a result, we don't add such indexes into the `ysql_ddl_txn_undergoing_subtransaction_rollback_map_[txn].tables` list. As a result, when the index is deleted as part of the base table deletion, the function `IsTableDeletionDueToRollbackToSubTxn` returns false for the index table. Therefore, we don't populate the txn id in the DeleteTable request leading to all transactions getting aborted.

The fix is to remember the index id when we skip its deletion in a map. The IsTableDeletionDueToRollbackToSubTxn function will use the map for such
indexes.

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestDdlSavepoints'

Reviewers: bkolagani, myang, #db-approvers

Reviewed By: bkolagani, #db-approvers

Subscribers: svc_phabricator, ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D48507