[#15507] docdb: Fix use-after-free post shutdown in cdc state table
Summary:
Currently there is a possible use-after-free issue in the cdc state table code. The cdc state table calls an async API and waits for the call to complete, or until the cdc service is shutdown. The callback the cdc state table passes to the async API uses cdc state table fields which are freed by the cdc service shutdown code. So if this callback is invoked after the cdc service has been shutdown, a use-after-free occurs. This use-after-free was introduced by 470e2df18787ca044a5fe7cb31e2c6306c09d401.

This diff wraps the state in a shared pointer and copies this shared pointer into the callback, keeping the state alive for the lifetime of the callback.

Test Plan:
```
./yb_build.sh asan --cxx-test cdcsdk_ysql-test --cxx-test-filter-re cdcsdk_ysql-test --gtest_filter 'CDCSDKYsqlTest.TestCDCSDKLagMetrics' -n 40 --stop-at-failure
```

Reviewers: devansh.singhal, bkolagani, sumukh.phalgaonkar

Reviewed By: sumukh.phalgaonkar

Subscribers: ybase, rthallam

Differential Revision: https://phorge.dev.yugabyte.com/D48599