[#29071] YSQL: SAOP merge pinning
Summary:
Commit 30bee61df303e81e5e97b918222c4d20fe9580f9 introduces SAOP merge as
an incomplete feature.  One incomplete part is that, given multiple SAOP
clauses for a key column, the one with lowest cardinality is chosen by
the planner, but that is not necessarily reflected in the executor.
Hook up the logic to use that pinned SAOP in the executor.  This helps
avoid the worst case situation where one SAOP has extremely low
cardinality and another has extremely high cardinality that surpasses
the yb_max_saop_merge_streams limit.

A major annoyance in getting this to work is handling the recheck code.
So what should be a simple change ends up touching a lot of areas.

Test Plan:
On Almalinux 8:

    ./yb_build.sh fastdebug --gcc13 daemons initdb --skip-pg-parquet \
      --java-test TestPgRegressSaopMerge

Backport-to: 2025.2

Reviewers: amartsinchyk

Reviewed By: amartsinchyk

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D47980