[PLAT-19116] Create Universe fails due to DeadlineExceeded being wrapped as INTERNAL
Summary: The original error was rethrown as INTERNAL. So, the change https://phorge.dev.yugabyte.com/D48608 was supposed to fix the error, but it caused it.

Test Plan:
UT to make sure it works.

Tried a few times locally. Manually created a 4 node universe, stop/start processes ran fine.

Reviewers: dshubin, amindrov

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D48702