[#27613] Docdb: Avoid aborting active transactions during DDLs if table locks are enabled.
Summary:
 With table locks enabled, we do not need to abort concurrent transactions when a DDL
is run.

 However, if there are any tservers that have not enabled table locks, we could have
a mixed mode where some transactions are using table locks while others do not.

 Two transactions that both respect and acquire table locks will not need to cancel
each other. However if either of them are not using table locks, we should use the old
behavior and abort concurrent transactions to be safe.

 We add a bool field to the TransactionMetadata to track whether or not the transaction
uses table locks. This is used in the StopActiveTransactionsPriorTo function to abort
the required transactions as per the updated rule above.

**Upgrade/Rollback safety:**
Should be ok.

Table locks will be enabled only after the upgrade is complete. The added fields will only be used if table locks are enabled to avoid aborting transactions that respect table locks.
If the release is rollbacked/downgraded, we will fall back on the old behavior which will abort all active transactions at a tablet.

Test Plan:
./yb_build.sh release --cxx-test pg_object_locks-test --gtest_filter TableLocksEnabled/PgObjectLocksTestAbortTxns.TestDDLAbortsTxns/*

./yb_build.sh release --cxx-test pg_object_locks-test --gtest_filter TableLocksEnabled/PgObjectLocksTestAbortTxns.TestDDLAbortsTxnsInMixedMode/*

./yb_build.sh release --cxx-test
conflict_resolve_keys_verification-itest

Reviewers: bkolagani, rthallam, pjain

Reviewed By: bkolagani, pjain

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D48753