[#29445] docdb: Vector index search hangs when the base table is split before vector index creation
Summary:
Vector index search does not work from the same session in which the vector index was created.
An attempt to run a search becomes stuck for the entire default timeout period during a tablet
lookup operation initiated in the MetaCache. The execution enters an infinite loop due to a
`ClientErrorCode::kTablePartitionListIsStale` status thrown from
`MetaCache::ProcessTabletLocation()`.

The error occurs because the vector index table’s partition list version does not match the
partition list version of the indexable table (it is lower). The check is performed for all tables
colocated on the same tablet during processing of each tablet returned by the lookup-table request,
and the vector index table is the only type of colocated table that can encounter this scenario,
because it may be created for a base table that had already been split.

The root cause is that the partition list version is not initialized during vector index table
creation: it is set to the initial value `0` instead of being initialized to the indexable table’s
partition list version, which corresponds to the actual partitions/tablets of the vector index
table.

However, even after fixing the hang, the search results from the vector index may still be
incorrect, because they are always taken from the lowest partition. This issue is tracked in:
https://github.com/yugabyte/yugabyte-db/issues/29543.

Test Plan:
yb_build.sh --cxx-test pgwrapper_pg_vector_index-test --gtest_filter PgDistributedVectorIndexTest.BaseTableManualSplitSimple/*

Manual test:
```
1. Create cluster with RF=1
2. =# CREATE EXTENSION vector;
3. =# CREATE TABLE test (id bigserial PRIMARY KEY, embedding vector(1)) split into 1 tablets;
4. =# INSERT INTO test SELECT i, ARRAY[i] FROM generate_series(0, 19) as i;
5. Flush the only tablet
6. Split the only tablet
7. Wait enough time (for example till parent tablet cleanup)
8. =# CREATE INDEX ON test USING ybhnsw (embedding vector_l2_ops);
9. Try search: =# SELECT * FROM test ORDER BY embedding <-> '[10.0]' limit 5;
10. Try to delete the table: =# DROP TABLE test;
```

Reviewers: sergei, rthallam

Reviewed By: sergei, rthallam

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D48739