[BACKPORT 2025.2][#29041] YSQL: Check request boundaries when building hash permutation
Summary:
Original commit: 0e094d25fe3ccfac7de107b793979f3bc57e5ac9 / D47730
We have an optimization for hash partitioned tables where we convert IN
and equality conditions on the key columns into a set of simple equality
conditions. However, that optimization did not take into account
possible request bounds.

Currently hash relation scan bounds can be set by a condition on
`yb_hash_code(<hash key>)` function. Other case that we about to enable
is parallel query. In both cases we want to use the request pruning feature
when we skip sending the requests where bounds make an empty range.
By current design request is skipped by setting its operation inactive.

This diff updates hash permutation logic to easier deal with the fact that
the template request may be out of bounds and hence permutation can be
skipped. That does not improve memory footprint a lot, since request has
to be created and boundaries set on it before it is checked. Main saving
come from not sending requests.

The pruning logic differs a bit in batching and non-batching mode.
In batching mode we lazily create one batch per partition. At the creation
time we combine existing request bounds with partition bounds and check
if the resulting range is empty. We don't create the IN condition in inactive
batches. That works as a criteria, if permutation's target batch exists, but does
not have the IN condition, we know it is inactive and skip the permutation.
We don't check if individual permutations are in bounds of their respective batch.
That may be a subject for future optimizations.

In non-batching mode we create one request per permutation. For each
permutation we use permutation's hash code to make the bounds of the
request and combine it with the existing. It may be a future optimization
too, to use hash values, and present range values to make narrower bounds.

Jira: DB-18798

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressHashInQueries#schedule'

Reviewers: jason, aagrawal

Reviewed By: jason

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D48749