[#29559] DocDB: Disable tablet split when table is being/was already hidden
Summary:
In one of the stress tests, we saw that a DDL's state maintained by `ysql_ddl_txn_verfication_state_map_` is never cleaned up due to processing of a split in parallel to a drop table when the database has a snapshot schedule enabled.

Without object locking, we observe the following consequences:
1. At the master, the table is left in a partially hidden state. i.e. `SysTablesEntryPB_HideState_HIDING` (some tablets in running state, and the rest in hidden state).
2. However, the tservers hide all the tablets including the new split ones. This is because the master receives heartbeats for the split tablets, and sees that the `hide_state` of the table as `HIDING` and then sends the delete/hide table rpcs.
This diverged table/tablets state between the master and the tservers could lead to subsequent problems.

With object locking enabled, the consequences are much severe. Since the ddl verifier task wrongly assumes that the table deletion isn't complete (because the newly split tablets didn't undergo deletion), it doesn't trigger cleanup of the state, which leads to a lock leak issue and the locks associated with the ddl transaction never get cleaned up.

The existing code prevents tablet splits while a deletion of the table is in progress/has completed. This revision extends the logic to additionally disable splits when the table is being hidden/was already hidden, thus resolving the above issue. With the fix, since splits are prevented once a drop is initiated, the ddl verifier successfully cleans up the state once all delete tablet rpcs go through.

Test Plan:
Jenkins

Was able to repro the lock leak issue locally with the following test
```
TEST_F(PgTabletSplitTest, TestDDLAmidstSplitExitsGracefully) {
  // Create snapshot to prevent deletion and force table into hidden state.
  {
    auto proxy = ASSERT_RESULT(cluster_->GetLeaderMasterProxy<master::MasterBackupProxy>());
    master::CreateSnapshotScheduleRequestPB req;
    auto table_identifier = req.mutable_options()->mutable_filter()->mutable_tables()->add_tables();
    table_identifier->mutable_namespace_()->set_name("yugabyte");
    table_identifier->mutable_namespace_()->set_database_type(YQLDatabase::YQL_DATABASE_PGSQL);
    req.mutable_options()->set_interval_sec(2);
    req.mutable_options()->set_retention_duration_sec(5);
    rpc::RpcController rpc;
    master::CreateSnapshotScheduleResponsePB resp;
    ASSERT_OK(proxy.CreateSnapshotSchedule(req, &resp, &rpc));
    ASSERT_OK(ResponseStatus(resp));
  }

  auto conn = ASSERT_RESULT(Connect());
  ASSERT_OK(conn.Execute("CREATE TABLE t(k INT PRIMARY KEY, v INT) SPLIT INTO 1 TABLETS"));
  ASSERT_OK(conn.Execute("INSERT INTO t SELECT i,i FROM generate_series(1, 10000) as i"));
  ASSERT_OK(conn.Execute("CREATE INDEX idx on t(v) SPLIT INTO 1 TABLETS"));
  ASSERT_OK(cluster_->FlushTablets());

  auto table_id = ASSERT_RESULT(GetTableIDFromTableName("idx"));
  const auto peers = ListTableActiveTabletLeadersPeers(cluster_.get(), table_id);
  ASSERT_EQ(1, peers.size());
  const auto& parent_peer = peers[0];
  ASSERT_OK(WaitForAnySstFiles(parent_peer));

  yb::SyncPoint::GetInstance()->LoadDependency({
    {
      "CatalogManager::DeleteOrHideTabletsAndSendRequests::Start",
      "TestDDLAmidstSplitExitsGracefully"
    },
    {
      "Tabletsplit::AddedChildrenTabletStateTableEntries",
      "CatalogManager::DeleteOrHideTabletsAndSendRequests::AddToMaps"
    }});

  yb::SyncPoint::GetInstance()->ClearTrace();
  yb::SyncPoint::GetInstance()->EnableProcessing();
  ANNOTATE_UNPROTECTED_WRITE(FLAGS_ddl_verification_timeout_multiplier) = 0;
  auto status_future = std::async(std::launch::async, [&]() -> Status {
    auto conn = VERIFY_RESULT(Connect());
    return conn.Execute("begin;DROP INDEX idx;commit");
  });
  TEST_SYNC_POINT("TestDDLAmidstSplitExitsGracefully");
  ASSERT_OK(InvokeSplitsAndWaitForDataCompacted(
      table_id, TabletSelector(1, SelectFirstTabletPolicy())));
  ASSERT_OK(status_future.get());

  ASSERT_OK(conn.Execute("SET statement_timeout='10s';"));
  ASSERT_OK(conn.Execute("ALTER TABLE t ADD COLUMN v1 INT DEFAULT 0"));
  ASSERT_OK(conn.Execute("INSERT INTO t VALUES(20000, 20000)"));
  ASSERT_OK(conn.Execute("CREATE INDEX idx on t(v) SPLIT INTO 1 TABLETS"));
}
```

Since the above test doesn't make sense with the fix, added a new test to assert that splits issues in concurrent to table being hidden are rejected.
```
./yb_build.sh --cxx-test='TEST_F(PgTabletSplitTest, TestDisableSplitWhenTableIsBeingHidden) {'
```

Reviewers: amitanand, rthallam, asrivastava

Reviewed By: asrivastava

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D48680